api/cli: expose rolling plan helper and shared solver hooks
